+++ backend/apps/weather/src/infrastructure/metrics/index.ts
@@ -1,1 +1,5 @@
-export * from "./metrics.factory";
+export * from "./http-metrics.service";
+export * from "./cache-metrics.service";
+export * from "./weather-provider-metrics.service";
+export * from "./subscription-metrics.service";
+export * from "./registry.manager";
