--- backend/apps/weather/src/infrastructure/metrics/metrics.factory.ts
@@ -1,8 +1,0 @@
-import { IMetricsService } from "@/shared/ports";
-import { MetricsService } from "./metrics.service";
-
-export class MetricsFactory {
-  static create(): IMetricsService {
-    return MetricsService.getInstance();
-  }
-}
